pyproject.toml
[tool.poetry]
name = "researcher"
version = "0.1.0"
description = "Simple Chat My Data for Youtube Videos"
authors = ["Oscar Laird <olaird25@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tqdm = "^4.65.0"
openai = "^0.27.7"
pinecone-client = "^2.2.1"
SQLAlchemy = "^2.0.15"
python-dotenv = "^1.0.0"
newspaper3k = "^0.2.8"
prettytable = "^3.7.0"


[tool.poetry.group.dev.dependencies]
fire = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
